requirements.txt
chromadb>=1.4.0
sentence-transformers>=5.0.0
torch>=2.0.0
groq>=1.0.0
numpy>=1.24.0
tqdm>=4.66.0
python-dotenv>=1.0.0
fastapi
uvicorn
python-multipart
pypdf
python-docx
streamlit
requests
